test(cypress): split APIM tests and platform tests
