+++ gravitee-apim-cypress/cypress/integration/apim/bulk/cypress-config.json
@@ -1,6 +1,6 @@
 {
-  "projectId": "gravitee-apim-cypress-e2e",
-  "integrationFolder": "cypress/integration/e2e",
+  "projectId": "gravitee-apim-cypress-bulk",
+  "integrationFolder": "cypress/integration/apim/bulk",
   "baseUrl": "http://localhost:8083",
   "video": false,
   "screenshotOnRunFailure": false,
