+++ gravitee-apim-cypress/cypress/integration/apim/bulk/ratings.test.ts
