+++ gravitee-apim-cypress/cypress/integration/apim/cypress-config.json
@@ -1,6 +1,6 @@
 {
-  "projectId": "gravitee-apim-cypress-bulk",
-  "integrationFolder": "cypress/integration/bulk",
+  "projectId": "gravitee-apim-cypress-apim",
+  "integrationFolder": "cypress/integration/apim",
   "baseUrl": "http://localhost:8083",
   "video": false,
   "screenshotOnRunFailure": false,
