+++ gravitee-apim-cypress/cypress/integration/apim/e2e/cypress-config.json
@@ -1,10 +1,10 @@
 {
-  "projectId": "gravitee-platform-cypress-rest-api",
-  "integrationFolder": "cypress/integration/rest-api",
-  "testFiles": "**/*.ts",
-  "baseUrl": "https://apim.gravitee.io/api",
+  "projectId": "gravitee-apim-cypress-e2e",
+  "integrationFolder": "cypress/integration/apim/e2e",
+  "baseUrl": "http://localhost:8083",
   "video": false,
   "screenshotOnRunFailure": false,
+  "testFiles": "**/*.ts",
   "env": {
     "failOnStatusCode": false,
     "api_publisher_user_login": "api1",
@@ -15,14 +15,10 @@
     "low_permission_user_password": "password",
     "admin_user_login": "admin",
     "admin_user_password": "admin",
-    "am_admin_user_login": "admin",
-    "am_admin_user_password": "adminadmin",
     "managementOrganizationApi": "/management/organizations/DEFAULT",
     "managementApi": "/management/organizations/DEFAULT/environments/DEFAULT",
-    "managementUI": "https://apim.gravitee.io/console",
-    "gatewayServer": "https://apim.gravitee.io",
-    "am_gatewayServer": "https://am.gravitee.io",
-    "am_managementAPI": "https://am.gravitee.io/management/organizations/DEFAULT/environments/DEFAULT",
+    "managementUI": "http://localhost:8084",
+    "gatewayServer": "http://localhost:8082",
     "portalApi": "/portal/environments/DEFAULT"
   }
 }
