+++ gravitee-apim-cypress/cypress/integration/apim/rest-api/apis/api-flow.ts
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { ADMIN_USER, LOW_PERMISSION_USER } from 'fixtures/fakers/users/users';
-import { ApiFakers } from 'fixtures/fakers/apis';
-import { PlanFakers } from 'fixtures/fakers/plans';
-import { createApi, deleteApi, publishApi, startApi, stopApi } from 'commands/management/api-management-commands';
-import { createPlan, publishPlan, deletePlan } from 'commands/management/api-plan-management-commands';
-import { Api } from 'model/apis';
-import { NewPlanEntity, PlanSecurityType } from 'model/plan';
+import { ADMIN_USER, LOW_PERMISSION_USER } from '@fakers/users/users';
+import { ApiFakers } from '@fakers/apis';
+import { PlanFakers } from '@fakers/plans';
+import { createApi, deleteApi, publishApi, startApi, stopApi } from '@commands/management/api-management-commands';
+import { createPlan, publishPlan, deletePlan } from '@commands/management/api-plan-management-commands';
+import { Api } from '@model/apis';
+import { NewPlanEntity, PlanSecurityType } from '@model/plan';
 import { requestGateway } from 'support/common/http.commands';
 
 context('Create an API flow', () => {
