+++ gravitee-apim-cypress/cypress/integration/apim/rest-api/apis/api-key-tests.ts
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 const jwt = require('jsonwebtoken');
-import { ADMIN_USER, API_PUBLISHER_USER } from 'fixtures/fakers/users/users';
-import { deleteApi, deployApi, getApiKeys, importCreateApi, startApi, stopApi } from 'commands/management/api-management-commands';
-import { publishPlan, closePlan } from 'commands/management/api-plan-management-commands';
-import { PlanSecurityType } from 'model/plan';
-import { ApiImportFakers } from 'fixtures/fakers/api-imports';
+import { ADMIN_USER, API_PUBLISHER_USER } from '@fakers/users/users';
+import { deleteApi, deployApi, getApiKeys, importCreateApi, startApi, stopApi } from '@commands/management/api-management-commands';
+import { publishPlan, closePlan } from '@commands/management/api-plan-management-commands';
+import { PlanSecurityType } from '@model/plan';
+import { ApiImportFakers } from '@fakers/api-imports';
 import { ApiImport } from '@model/api-imports';
 import {
   closeApplicationSubscription,
