+++ gravitee-apim-cypress/cypress/integration/apim/rest-api/cypress-config.json
@@ -1,0 +1,24 @@
+{
+  "projectId": "gravitee-apim-cypress-rest-api",
+  "integrationFolder": "cypress/integration/apim/rest-api",
+  "baseUrl": "http://localhost:8083",
+  "video": false,
+  "screenshotOnRunFailure": false,
+  "testFiles": "**/*.ts",
+  "env": {
+    "failOnStatusCode": false,
+    "api_publisher_user_login": "api1",
+    "api_publisher_user_password": "api1",
+    "application_user_login": "application1",
+    "application_user_password": "application1",
+    "low_permission_user_login": "user",
+    "low_permission_user_password": "password",
+    "admin_user_login": "admin",
+    "admin_user_password": "admin",
+    "managementOrganizationApi": "/management/organizations/DEFAULT",
+    "managementApi": "/management/organizations/DEFAULT/environments/DEFAULT",
+    "managementUI": "http://localhost:8084",
+    "gatewayServer": "http://localhost:8082",
+    "portalApi": "/portal/environments/DEFAULT"
+  }
+}
