+++ gravitee-apim-cypress/cypress/integration/cypress-config.json
@@ -1,0 +1,28 @@
+{
+  "projectId": "gravitee-apim-cypress",
+  "integrationFolder": "cypress/integration",
+  "testFiles": "**/*.ts",
+  "baseUrl": "https://apim.gravitee.io/api",
+  "video": false,
+  "screenshotOnRunFailure": false,
+  "env": {
+    "failOnStatusCode": false,
+    "api_publisher_user_login": "api1",
+    "api_publisher_user_password": "api1",
+    "application_user_login": "application1",
+    "application_user_password": "application1",
+    "low_permission_user_login": "user",
+    "low_permission_user_password": "password",
+    "admin_user_login": "admin",
+    "admin_user_password": "admin",
+    "am_admin_user_login": "admin",
+    "am_admin_user_password": "adminadmin",
+    "managementOrganizationApi": "/management/organizations/DEFAULT",
+    "managementApi": "/management/organizations/DEFAULT/environments/DEFAULT",
+    "managementUI": "https://apim.gravitee.io/console",
+    "gatewayServer": "https://apim.gravitee.io",
+    "am_gatewayServer": "https://am.gravitee.io",
+    "am_managementAPI": "https://am.gravitee.io/management/organizations/DEFAULT/environments/DEFAULT",
+    "portalApi": "/portal/environments/DEFAULT"
+  }
+}
