+++ gravitee-apim-cypress/cypress/integration/platform/oauth2-policy-tests.ts
