+++ gravitee-apim-cypress/package.json
@@ -4,14 +4,15 @@
     "private": true,
     "description": "Gravitee.io APIM - Cypress",
     "scripts": {
-        "test": "cypress open --config-file cypress/integration/rest-api/platform_cypress-config.json",
-        "test:bulk": "cypress run --config-file cypress/integration/bulk/cypress-config.json",
-        "test:rest-api": "cypress run --config-file cypress/integration/rest-api/cypress-config.json",
-        "test:e2e": "cypress open --config-file cypress/integration/e2e/cypress-config.json",
-        "test:ui": "cypress open --config-file cypress/integration/ui/cypress-config.json",
-        "test:dev": "cypress run --env printLogsToConsole=always --config-file cypress/integration/rest-api/cypress-config.json",
-        "test:parallel": "cypress run --record --group apis --spec 'cypress/integration/rest-api/apis/**/*' --config-file cypress/integration/rest-api/cypress-config.json",
-        "test:platform:rest-api": "cypress open --config-file cypress/integration/rest-api/platform_cypress-config.json",
+        "test": "cypress open --config-file cypress/integration/cypress-config.json",
+        "test:apim": "cypress open --config-file cypress/integration/apim/cypress-config.json",
+        "test:apim:bulk": "cypress run --config-file cypress/integration/apim/bulk/cypress-config.json",
+        "test:apim:rest-api": "cypress run --config-file cypress/integration/apim/rest-api/cypress-config.json",
+        "test:apim:e2e": "cypress open --config-file cypress/integration/apim/e2e/cypress-config.json",
+        "test:apim:ui": "cypress open --config-file cypress/integration/apim/ui/cypress-config.json",
+        "test:apim:dev": "cypress run --env printLogsToConsole=always --config-file cypress/integration/apim/rest-api/cypress-config.json",
+        "test:apim:parallel": "cypress run --record --group apis --spec 'cypress/integration/apim/rest-api/apis/**/*' --config-file cypress/integration/apim/rest-api/cypress-config.json",
+        "test:platform": "cypress open --config-file cypress/integration/platform/cypress-config.json",
         "lint": "echo 'Lint to be configured'",
         "lint:fix": "npm run prettier:fix && npm run lint:license:fix",
         "lint:license": "license-check-and-add check -f license-check-config.json",
